pyproject.toml
[project]
name = "llmstart-homework-ilya"
version = "0.1.0"
description = "LLM-ассистент для консультаций в Telegram"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 